security fix from NKJG (a developer of JP version Nucleus): delete draft only if the member has the right to alter this item
git-svn-id: https://nucleuscms.svn.sourceforge.net/svnroot/nucleuscms/trunk@1188 dbb57b54-ea0f-0410-8215-9b63e7fa3c7b

+++ nucleus/nucleus/libs/ADMIN.php
@@ -1174,7 +1174,7 @@ class ADMIN {
 
 		$this->updateFuturePosted($blogid);
 
-		if ($draftid > 0) {
+		if ($draftid > 0 && $member->canAlterItem($draftid)) {
 			ITEM::delete($draftid);
 		}
